Work arround bug with server communication.
It works, but is an ugly hack : sometimes server doesn't
respond to connection (and so there's no getHeaderField(0)
and getResponseCode give us an IOException), so if we try to connect
again it seems to work. It would be better to find why server
doesn't respond (don't know if it's our issue or an rssLounge issue).